feat(onErrorResumeNext): add higher-order lettable version of onErrorResumeNext
